Fix data sizes used in Itdb_ArtworkFormat
The mostly automated SysInfoExtended parser expects the fields in the
Itdb_ArtworkFormat structure to be of type gint. They used to be of type gin
which led to random overwriting of their value since the parser expected the
to be bigger than they really are.

git-svn-id: https://gtkpod.svn.sourceforge.net/svnroot/gtkpod/libgpod/trunk@2139 f01d2545-417e-4e96-918e-98f8d0dbbcb6

+++ src/itdb_device.h
@@ -95,9 +95,9 @@ struct _Itdb_Device
 };
 
 struct _Itdb_ArtworkFormat {
-        gint16 format_id;
-        gint16 width;
-        gint16 height;
+        gint format_id;
+        gint width;
+        gint height;
         ItdbThumbFormat format;
         gint32 padding; /* not found in SysInfoExtended, added  
                          * for compatibility with hardcoded artwork formats */
